Draw trains from scratch every frame again
Since the position of trains changes every frame anyway,
I don't think there is any point in trying to batch the
draw calls as long as there is only a single shape per
Train object.

When there are several shapes per Train object however,
then each Train should be a ShapeList which we can move
and rotate all together.

All this should of course be encapsulated inside a
Drawer class or similar.